chore: adjust platform admin filter logic (#4494)
