+++ editor.planx.uk/src/pages/PlatformAdminPanel/PlatformAdminPanel.test.tsx
@@ -6,7 +6,6 @@ import { mockTeams } from "ui/shared/DataTable/mockTeams";
 import { it } from "vitest";
 
 import { PlatformAdminPanel } from "./PlatformAdminPanel";
-import { internalTeamNames } from "./utils";
 
 const { getState, setState } = useStore;
 
@@ -51,14 +50,6 @@ describe("Platform admin panel", () => {
     expect(
       screen.getAllByText("Apply for a lawful development certificate"),
     ).toHaveLength(2); // Two teams in the mock data have an LDC flow
-  });
-
-  it("does not show rows for internal teams", () => {
-    setup(<PlatformAdminPanel />);
-
-    internalTeamNames.map((internalTeam) =>
-      expect(screen.queryByText(internalTeam)).not.toBeInTheDocument(),
-    );
   });
 
   it("renders a tick / cross for boolean values", () => {
